requirements.txt
open virtual environment

pip install:
  django==3.1.3
  djangorestframework==3.12.2
  python==3.9.0


